Sample DB Data
Adding scripts to populate our temporary DB. This allows us to really
get down and dirty with our pitches. They are denoted with file names
beginning with X.Y where X is the number of the main table data will be
added to, and Y is the order to run these additions in.